% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{dat_fmtfun}
\alias{dat_fmtfun}
\title{Function including factor and variable labels}
\format{
function
}
\description{
A function to apply factor and variable labels to user input data.
}
\keyword{datasets}
