% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{patdat_default}
\alias{patdat_default}
\title{Default user input values}
\format{
data frame
}
\description{
A data frame including the default values for user inputs. Default values were taken as the mode or median value for each item.
}
\keyword{datasets}
