% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{xtest_vnames}
\alias{xtest_vnames}
\title{Variables to pass to predictive models}
\format{
vector
}
\description{
A vector of the variable names that the predictive models were trained on.
}
\keyword{datasets}
