% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{zcta_hhinc}
\alias{zcta_hhinc}
\title{Zip-code median household income look-up table}
\format{
data frame with 1 row per zip-code
}
\description{
A data frame of median household income by zip-code, obtained from the 2019 ACS.
}
\keyword{datasets}
